application/requirements.txt
# Core Framework
streamlit==1.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Database Connectivity
sqlalchemy==2.0.25
pymysql==1.1.0
pyodbc==5.0.1

# Data Visualization
plotly==5.18.0
altair==4.2.2
matplotlib==3.8.2
seaborn==0.13.1

# Excel/CSV Handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1

# Data Quality
great-expectations==0.18.8

# Authentication (Optional)
streamlit-authenticator==0.2.3

# Additional UI Components
streamlit-extras==0.3.6
streamlit-option-menu==0.3.6

requests==2.31.0
python-dotenv==1.0.1


mysql-connector-python==8.1.0
pymysql==1.1.0
pillow==10.0.0
openpyxl==3.1.2
redis==5.0.0


